Refs #561: Create StudentGradeAssembler class + tests + interface
Changed attribute data type in StudentGradeRequestDTO

+++ src/main/java/PAI/dto/studentGrade/StudentGradeRequestDTO.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.Positive;
 public record StudentGradeRequestDTO(
 
         @Positive(message = "Grade must be a positive number.")
-        int grade,
+        double grade,
 
         @NotBlank(message = "Date is required.")
         String date,
